Add Status returns, FileSourceBase/StreamSourceBase, manifest schema
SDK virtuals (bindWriteHost, bindRuntimeHost, loadConfig, start, pause,
resume, poll) now return PJ::Status instead of bool. Trampolines unwrap
Status into bool + last_error_ at the ABI boundary — no protocol bump.

New derived base classes FileSourceBase and StreamSourceBase
(data_source_patterns.hpp) encapsulate the lifecycle state machine for
the two dominant plugin patterns, exposing a narrower virtual surface
(importData / onStart+onPoll+onStop).

Document file_extensions array in the manifest JSON schema so the host
can build file-dialog filters without instantiating plugins.

Rewrite data-source-guide.md Common Patterns with the new base classes,
Status error handling, and manifest schema reference.

+++ pj_datastore/src/plugin_data_host.cpp
@@ -756,11 +756,12 @@ struct ToolboxCore {
                 is_null ? 0 : decodeNumericExact<uint64_t>(chunk, static_cast<std::size_t>(col_index), row));
             break;
           case PrimitiveType::kBool:
-            state->bool_values.push_back(is_null ? 0 : static_cast<uint8_t>(chunk.readBool(col_index, row)));
+            state->bool_values.push_back(
+                is_null ? 0 : static_cast<uint8_t>(chunk.readBool(static_cast<std::size_t>(col_index), row)));
             break;
           case PrimitiveType::kString: {
             if (!is_null) {
-              const auto text = chunk.readString(col_index, row);
+              const auto text = chunk.readString(static_cast<std::size_t>(col_index), row);
               state->string_bytes.insert(state->string_bytes.end(), text.begin(), text.end());
             }
             state->string_offsets.push_back(static_cast<uint32_t>(state->string_bytes.size()));
